backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
psycopg[binary]>=3.2
django-environ>=0.11
channels>=4.0
channels-redis>=4.2
daphne>=4.1
drf-spectacular>=0.27
pywebpush>=1.14
bleach>=6.1

dj-database-url>=2.2
